Revert "Most viewed container on fronts"
