+++ facia/app/AppLoader.scala
@@ -1,4 +1,3 @@
-import agents.MostViewedAgent
 import akka.actor.ActorSystem
 import app.{FrontendApplicationLoader, FrontendComponents}
 import com.softwaremill.macwire._
@@ -8,10 +7,8 @@ import common.dfp.FaciaDfpAgentLifecycle
 import concurrent.BlockingOperations
 import conf.switches.SwitchboardLifecycle
 import conf.CachedHealthCheckLifeCycle
-import contentapi.{CapiHttpClient, ContentApiClient, HttpClient}
 import controllers.{FaciaControllers, HealthCheck}
 import dev.{DevAssetsController, DevParametersHttpRequestHandler}
-import feed.MostViewedLifecycle
 import http.{CommonFilters, PreloadFilters}
 import model.ApplicationIdentity
 import services.ophan.SurgingContentAgentLifecycle
@@ -40,13 +37,10 @@ trait FapiServices {
 
 trait AppComponents extends FrontendComponents with FaciaControllers with FapiServices {
 
-  lazy val capiHttpClient: HttpClient = wire[CapiHttpClient]
-  lazy val contentApiClient = wire[ContentApiClient]
   lazy val healthCheck = wire[HealthCheck]
   lazy val devAssetsController = wire[DevAssetsController]
   lazy val ophanApi = wire[OphanApi]
   lazy val logbackOperationsPool = wire[LogbackOperationsPool]
-  lazy val mostViewedAgent = wire[MostViewedAgent]
 
   override lazy val lifecycleComponents = List(
     wire[LogstashLifecycle],
@@ -57,7 +51,6 @@ trait AppComponents extends FrontendComponents with FaciaControllers with FapiSe
     wire[IndexListingsLifecycle],
     wire[SwitchboardLifecycle],
     wire[CachedHealthCheckLifeCycle],
-    wire[MostViewedLifecycle],
   )
 
   lazy val router: Router = wire[Routes]
