+++ facia/app/controllers/FaciaControllers.scala
@@ -1,6 +1,5 @@
 package controllers
 
-import agents.MostViewedAgent
 import com.softwaremill.macwire._
 import model.ApplicationContext
 import play.api.libs.ws.WSClient
@@ -11,7 +10,6 @@ trait FaciaControllers {
   def frontJsonFapiLive: FrontJsonFapiLive
   def controllerComponents: ControllerComponents
   def wsClient: WSClient
-  def mostViewedAgent: MostViewedAgent
   implicit def appContext: ApplicationContext
   lazy val faciaController = wire[FaciaControllerImpl]
 }
