+++ facia/test/FaciaMetaDataTest.scala
@@ -1,6 +1,5 @@
 package metadata
 
-import agents.MostViewedAgent
 import akka.actor.ActorSystem
 import com.gu.facia.client.models.{ConfigJson, FrontJson}
 import concurrent.BlockingOperations
@@ -13,7 +12,7 @@ import org.scalatest.{BeforeAndAfterAll, DoNotDiscover}
 import org.scalatestplus.mockito.MockitoSugar
 import play.api.libs.json._
 import play.api.test.Helpers._
-import services.{ConfigAgent, OphanApi}
+import services.ConfigAgent
 
 import scala.concurrent.duration._
 import scala.concurrent.Await
@@ -28,8 +27,7 @@ import test._
     with WithMaterializer
     with WithTestWsClient
     with MockitoSugar
-    with WithTestFrontJsonFapi
-    with WithTestContentApiClient {
+    with WithTestFrontJsonFapi {
 
   override def beforeAll(): Unit = {
     val refresh = ConfigAgent.refreshWith(
@@ -42,12 +40,7 @@ import test._
     Await.result(refresh, 3.seconds)
   }
 
-  lazy val faciaController = new FaciaControllerImpl(
-    fapi,
-    play.api.test.Helpers.stubControllerComponents(),
-    wsClient,
-    new MostViewedAgent(testContentApiClient, new OphanApi(wsClient), wsClient),
-  )
+  lazy val faciaController = new FaciaControllerImpl(fapi, play.api.test.Helpers.stubControllerComponents(), wsClient)
   val frontPath = "music"
 
   it should "Include organisation metadata" in {
