+++ preview/app/AppLoader.scala
@@ -1,4 +1,4 @@
-import agents.{CuratedContentAgent, MostViewedAgent}
+import agents.CuratedContentAgent
 import akka.actor.ActorSystem
 import app.{FrontendApplicationLoader, FrontendComponents, LifecycleComponent}
 import com.softwaremill.macwire._
@@ -92,7 +92,6 @@ trait PreviewControllerComponents
   lazy val faviconController = wire[FaviconController]
   lazy val itemController = wire[ItemController]
   lazy val oAuthLoginController = wire[OAuthLoginPreviewController]
-  lazy val mostViewedAgent = wire[MostViewedAgent]
 }
 
 trait AppComponents
